Prepare to migrate to github pages for hosting
At present, zeespencer.com is hosted on an AWS S3 bucket; which works
pretty well but requires me to keep AWS Access Keys up to date and
figured out; which adds to the carry-cost of maintaining my personal
site; which makes it less likely for me to update it.

By switching to github-pages, I hope to minimize the complexity of
distributing the site while still maintaining a reasonable degree of
control of the build pipeline.