The site may be deployed again
I migrated off of a customer S3 bucket and Cloudfront domain as it is
significantly less complex to rely on github-pages. This does mean I'm
slightly trapped in their ecosystem, but migrating the deployment and
DNS configuration should be trivial when I do have the bandwidth to look
for non-github alternatives.