📝 First draft of choosing between web ui libaries
This particular article is inspired from conversations we've been having
in the coop about what UI libraries we want to use as we prototype
products and services.

My goal here is to share some of my values and preferences in a space
that isn't dictating terms for the cooperative; while also creating a
useful resource for folks who are learning how to make architectural
decisions in their own context.